requirements.txt
asgiref==3.7.2
Django==3.1.14
psycopg2-binary
python-dotenv==1.0.0
